+++ exonum/benches/criterion/block.rs
@@ -52,7 +52,7 @@ use futures::sync::mpsc;
 use rand::{rngs::StdRng, Rng, SeedableRng};
 use tempdir::TempDir;
 
-use std::iter;
+use std::{collections::BTreeMap, iter};
 
 /// Number of transactions added to the blockchain before the bench begins.
 const PREPARE_TRANSACTIONS: usize = 10_000;
@@ -85,7 +85,7 @@ fn create_blockchain(db: impl Database, services: Vec<Box<dyn Service>>) -> Bloc
     };
     use std::sync::Arc;
 
-    let dummy_channel = mpsc::unbounded();
+    let dummy_channel = mpsc::channel(0);
     let service_keypair = (PublicKey::zero(), SecretKey::zero());
     let mut blockchain = Blockchain::new(
         Arc::new(db) as Arc<dyn Database>,
@@ -106,7 +106,12 @@ fn create_blockchain(db: impl Database, services: Vec<Box<dyn Service>>) -> Bloc
 }
 
 fn execute_block(blockchain: &Blockchain, height: u64, txs: &[Hash]) -> (Hash, Patch) {
-    blockchain.create_patch(ValidatorId::zero(), Height(height), txs)
+    blockchain.create_patch(
+        ValidatorId::zero(),
+        Height(height),
+        txs,
+        &mut BTreeMap::new(),
+    )
 }
 
 mod timestamping {
@@ -474,7 +479,7 @@ fn prepare_blockchain(
         // We make use of the fact that `Blockchain::commit()` doesn't check
         // precommits in any way (they are checked beforehand by the consensus algorithm).
         blockchain
-            .commit(&patch, block_hash, iter::empty())
+            .commit(&patch, block_hash, iter::empty(), &mut BTreeMap::new())
             .unwrap();
     }
 }