+++ exonum/benches/criterion/transactions.rs
@@ -18,12 +18,7 @@ const SAMPLE_SIZE: usize = 20;
 use criterion::{
     AxisScale, Bencher, Criterion, ParameterizedBenchmark, PlotConfiguration, Throughput,
 };
-use futures::{
-    stream,
-    sync::mpsc::{Sender, UnboundedSender},
-    sync::oneshot,
-    Future, Sink,
-};
+use futures::{stream, sync::mpsc::Sender, sync::oneshot, Future, Sink};
 use tokio_core::reactor::Core;
 
 use std::{
@@ -134,7 +129,7 @@ struct MessageVerifier {
     tx_handler: MessagesHandlerRef,
     network_thread: JoinHandle<()>,
     handler_thread: JoinHandle<()>,
-    api_sender: Option<UnboundedSender<ExternalMessage>>,
+    api_sender: Option<Sender<ExternalMessage>>,
     network_sender: Option<Sender<NetworkEvent>>,
 }