+++ exonum/src/node/consensus.rs
@@ -14,7 +14,7 @@
 
 use std::collections::HashSet;
 
-use crate::blockchain::Schema;
+use crate::blockchain::{check_tx, Schema};
 use crate::crypto::{CryptoHash, Hash, PublicKey};
 use crate::events::InternalRequest;
 use crate::helpers::{Height, Round, ValidatorId};
@@ -122,7 +122,7 @@ impl NodeHandler {
         ) {
             Ok(state) => state.has_unknown_txs(),
             Err(err) => {
-                warn!("{}, msg={:?}", err, msg);
+                warn!("{} from {:?}", err, msg.validator());
                 return;
             }
         };
@@ -480,7 +480,12 @@ impl NodeHandler {
             // FIXME: Avoid of clone here. (ECR-171)
             let block_state = self.state.block(&block_hash).unwrap().clone();
             self.blockchain
-                .commit(block_state.patch(), block_hash, precommits)
+                .commit(
+                    &block_state.patch(),
+                    block_hash,
+                    precommits,
+                    self.state.tx_cache_mut(),
+                )
                 .unwrap();
             // Update node state.
             self.state
@@ -533,7 +538,9 @@ impl NodeHandler {
         let hash = msg.hash();
 
         let snapshot = self.blockchain.snapshot();
-        if Schema::new(&snapshot).transactions().contains(&hash) {
+        let schema = Schema::new(&snapshot);
+
+        if check_tx(&hash, &schema.transactions(), self.state.tx_cache()) {
             bail!("Received already processed transaction, hash {:?}", hash)
         }
 
@@ -542,14 +549,7 @@ impl NodeHandler {
             bail!("Received malicious transaction.")
         }
 
-        let fork = self.blockchain.fork();
-        {
-            let mut schema = Schema::new(&fork);
-            schema.add_transaction_into_pool(msg);
-        }
-        self.blockchain
-            .merge(fork.into_patch())
-            .expect("Unable to save transaction to persistent pool.");
+        self.state.tx_cache_mut().insert(hash, msg);
 
         if self.state.is_leader() && self.state.round() != Round::zero() {
             self.maybe_add_propose_timeout();
@@ -682,17 +682,10 @@ impl NodeHandler {
             if self.state.have_prevote(round) {
                 return;
             }
-            let snapshot = self.blockchain.snapshot();
-            let schema = Schema::new(&snapshot);
-            let pool = schema.transactions_pool();
-            let pool_len = schema.transactions_pool_len();
-
-            info!("LEADER: pool = {}", pool_len);
-
             let round = self.state.round();
-            let max_count = ::std::cmp::min(u64::from(self.txs_block_limit()), pool_len);
 
-            let txs: Vec<Hash> = pool.iter().take(max_count as usize).collect();
+            let txs = self.get_txs_for_propose();
+
             let propose = self.sign_message(Propose::new(
                 validator_id,
                 self.state.height(),
@@ -717,6 +710,33 @@ impl NodeHandler {
                 self.handle_majority_prevotes(round, &hash);
             }
         }
+    }
+
+    fn get_txs_for_propose(&self) -> Vec<Hash> {
+        let txs_cache_len = self.state.tx_cache_len() as u64;
+        let tx_block_limit = self.txs_block_limit();
+
+        let snapshot = self.blockchain.snapshot();
+        let schema = Schema::new(&snapshot);
+        let pool = schema.transactions_pool();
+        let pool_len = schema.transactions_pool_len();
+
+        info!("LEADER: pool = {}, cache = {}", pool_len, txs_cache_len);
+
+        let remaining_tx_count = tx_block_limit.saturating_sub(txs_cache_len as u32);
+        let cache_max_count = ::std::cmp::min(u64::from(tx_block_limit), txs_cache_len);
+
+        let mut cache_txs: Vec<Hash> = self
+            .state
+            .tx_cache()
+            .keys()
+            .take(cache_max_count as usize)
+            .cloned()
+            .collect();
+        let pool_txs: Vec<Hash> = pool.iter().take(remaining_tx_count as usize).collect();
+
+        cache_txs.extend(pool_txs);
+        cache_txs
     }
 
     /// Handles request timeout by sending the corresponding request message to a peer.
@@ -781,7 +801,12 @@ impl NodeHandler {
         height: Height,
         tx_hashes: &[Hash],
     ) -> (Hash, Patch) {
-        self.blockchain.create_patch(proposer_id, height, tx_hashes)
+        self.blockchain.create_patch(
+            proposer_id,
+            height,
+            tx_hashes,
+            &mut self.state.tx_cache_mut(),
+        )
     }
 
     /// Calls `create_block` with transactions from the corresponding `Propose` and returns the